refactor(filters): Split out date filters
